Add a helper function for connecting to Redshift.
This commit introduces a new `odbc::redshift()` function intended to
help setting up connections to AWS Redshift, especially when using IAM
credentials. It generally follows the pattern established with
`odbc::databricks()` and `odbc::snowflake()`.

Note that finding IAM credentials is outsourced to `paws.common`. Some
Redshift ODBC drivers (there are a few of them) can handle AWS profiles
or IAM role assumption, and I did have an earlier version of this commit
that did this work manually; unfortunately the supported parameters
depend not only on the driver, but also the driver version, and it
quickly became a mess.

So: using `paws.common` here allows us to be very driver
version-agnostic, and ensures we support as wide a number of IAM setups
as possible.

Note that I also refectored some of automatic driver discover for
Databricks and Snowflake into its own helper utility so we could use the
same logic for Redshift, and fixed some grammar issues in their
respective error messages.

Unit tests are included.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>